extract DelayList view from ContentView
